Refactor: Change "player" to "client"
